+++ crates/behavior/src/effects/mod.rs
@@ -1,7 +1,7 @@
 mod actions;
 mod sending;
 
-pub use actions::{Acted, Actions, Become};
+pub use actions::{Acted, Actions, AppendSend, Become};
 pub use sending::{
     InterpretDelivery, InterpretRequest, InterpretSends, InterpreterRequest, InterpreterRequests,
     NoReturnToEmitter, NoSends, Own, ReturnsToEmitter, SendEffects, SendInput, SendInterpreter,